// packages/somoto/src/types.ts
export type ToastTypes =
  | 'normal'
  | 'action'
  | 'success'
  | 'info'
  | 'warning'
  | 'error'
  | 'loading'
  | 'default';

export type PromiseT<Data = any> = Promise<Data> | (() => Promise<Data>);
// ...
export interface ToastClassnames {
  toast?: string;
  title?: string;
  description?: string;
  loader?: string;
  closeButton?: string;
  cancelButton?: string;
  actionButton?: string;
  success?: string;
  error?: string;
  info?: string;
  warning?: string;
  loading?: string;
  default?: string;
  content?: string;
  icon?: string;
}

export interface ToastIcons {
  success?: JSX.Element;
  info?: JSX.Element;
  warning?: JSX.Element;
  error?: JSX.Element;
  loading?: JSX.Element;
  close?: JSX.Element;
}

export interface Action {
  label: JSX.Element;
  onClick: JSX.EventHandler<HTMLButtonElement, MouseEvent>;
  actionButtonStyle?: JSX.CSSProperties;
}

export interface ToastType {
  id: number | string;
  title?: string | JSX.Element;
  type?: ToastTypes;
  icon?: JSX.Element;
  jsx?: JSX.Element;
  richColors?: boolean;
  invert?: boolean;
  closeButton?: boolean;
  dismissible?: boolean;
  description?: JSX.Element;
  duration?: number;
  delete?: boolean;
  important?: boolean;
  action?: Action | JSX.Element;
  cancel?: Action | JSX.Element;
  onDismiss?: (toast: ToastType) => void;
  onAutoClose?: (toast: ToastType) => void;
  promise?: PromiseT;
  cancelButtonStyle?: JSX.CSSProperties;
  actionButtonStyle?: JSX.CSSProperties;
  style?: JSX.CSSProperties;
  unstyled?: boolean;
  className?: string;
  classNames?: ToastClassnames;
  descriptionClassName?: string;
  position?: Position;
}

export function isAction(action: Action | JSX.Element): action is Action {
  return (action as Action).label !== undefined;
}

export type Position =
  | 'top-left'
  | 'top-right'
  | 'bottom-left'
  | 'bottom-right'
  | 'top-center'
  | 'bottom-center';
export interface HeightT {
  height: number;
  toastId: number | string;
  position: Position;
}

export interface ToastOptions {
  className?: string;
  closeButton?: boolean;
  descriptionClassName?: string;
  style?: JSX.CSSProperties;
  cancelButtonStyle?: JSX.CSSProperties;
  actionButtonStyle?: JSX.CSSProperties;
  duration?: number;
  unstyled?: boolean;
  classNames?: ToastClassnames;
}

type CnFunction = (...classes: Array<string | undefined>) => string;
// ...
export interface ToastProps {
  toast: ToastType;
  toasts: ToastType[];
  index: number;
  expanded: boolean;
  invert: boolean;
  heights: HeightT[];
  setHeights: Setter<HeightT[]>;
  removeToast: (toast: ToastType) => void;
  gap?: number;
  position: Position;
  visibleToasts: number;
  expandByDefault: boolean;
  closeButton: boolean;
  interacting: boolean;
  style?: JSX.CSSProperties;
  cancelButtonStyle?: JSX.CSSProperties;
  actionButtonStyle?: JSX.CSSProperties;
  duration?: number;
  class?: string;
  classNames?: ToastClassnames;
  unstyled?: boolean;
  descriptionClassName?: string;
  icons?: ToastIcons;
  closeButtonAriaLabel?: string;
  pauseWhenPageIsHidden: boolean;
  cn: CnFunction;
  defaultRichColors?: boolean;
}